comment out seobnr final frequency (#962)
The SEOBNR final frequency function doesn't seem to work properly. For some waveforms it seems to end far earlier than the model naturally does, and causes truncation within the ligo band of the waveform. This truncation causes the background to increase significantly.

For now, the only solution I have is to deactivate this. Functions such as the interpolated waveforms will still work as long as they have a time estimates.