Remove unnecessary and impossible automatic transforms  (#2435)
* Remove unnecessary and harmful transforms from the automatically applied transforms.

Add only one line in get_common_cbc_transforms:
Before adding a transform from the base parameters to the requested parameters to the list of transforms which will be executed (from_base_c), check if the intersection between requested parameters and output of the transform is already in the sampled parameters. If so, do NOT add the transform.

Applying these transforms here is not useful, since they only add parameters that are already known. Also, adding this would lead to errors for example in the case of Mass1Mass2ToMchirpQ:
If sampling is done in q and m_total and q is requested, then mass1 and mass2 are asked for by the transform and not known, so an error will occur.

* Fix style for code climate 

Is
`if not intersect`
better than
`if len(intersect) < 1` ? 
To me the first formulation suggested by code climate seems less clear,
and also continues to run if nonsense is saved in intersect (ie if it is not a list).