Add base class for nested samplers
For now, this mainly adds generic checkpointing
facility that can be inherited by individual
sampler classes. The goal here is to enable a
fall-back checkpointing option for new and
recently added samplers, with little development
overhead.

Algo to checkpoint:

- Write out samples and data to regular inference.hdf.*
- Strip un-picklable attributes of sampler object
- Pickle sampler state

Algo to resume:

- Initialize pycbc's sampler class
- Deserealize the external sampler's object
- Set attributes in it that were not
  serializable
- Continue sampling

On inheriting this base class, individual sampler
classes in pycbc will need to provide methods to:

 - strip un-serializable attributes for checkpointing
 - set the same attributes appropriately when resuming

using information stored to disk. Current triggers of
checkpoint are

 - user-provided checkpoint-interval
 - user-provided checkpoint-duration
 - interrupting signals (for e.g., c.f. job schedulers)