incorporated the necessary changes into it single_template_p
added num_sample which when None, will require 'polarization' as a parameter. 
If num_sample has some int value, an array is created, over which marginalization takes place.
If num_sample has int value, and polarization is given as a param, then array is still created and marginalization is done.