Changed the gating to use the default method of the time series. (#3673)
* Changed the gating to use the default method of the time series.

Added a new option --autogating-method to choose between different
available methods to apply the gates.

* -Changed the option --autogating-method to --gating-method.
-Changed how the gates are applied not only for auto-gates but
 provided ones as well.
-Adjusted the encoding of strings in the git-info of setup.py to
 allow for umlauts.

* Added encode and decode options to the GitInfo object.

Calling this function asserts that all strings have the same
encoding.
The default error-behavior for the encoding is set to `replace`,
such that unknown characters don't break the function but get
replaced.

* Encoding strings that are written to the file.

Removed the encode and decode funtion from GitInfo, as they are not
required.
Encoding every string that is written to setup.py. Changed the
write-mode of the file to be bytes.

* Fixed wrong name of opt.

Made some changes to please codeclimate.