Update pycbc_submit_dax to fix early proxy expiration on ./start resubmissions. (#3755)
* Update pycbc_submit_dax

Adding the -hours argument to grid-proxy-init in the case you do not have a RFC compliant proxy.  This prevents a failure mode where your workflow runs longer than 12 hours after you resubmit it and then all the jobs start to fail due to expired proxies.  276 hours is equal to the maximum standard proxy lifetime of 11.5 days that is created when you use ligo-proxy-init.

* Update pycbc_submit_dax

Co-authored-by: Ian Harry <iwharry@googlemail.com>