Switch to minimal pegasus api requirement
This is an attempt to address the immediate issue of Pegasus having a hard requirement on flask which no longer works due to #3942. In general, it's probably a good idea to keep our dependencies as lean as possible as well.

Assuming this passes tests, is there any reason we don't *just* depend on the API? Doe we actually need anything else for the PyCBC library directly?