Change to start GraceDB uploads and setup_optimize_snr process in threads
These changes allow the Rank0 MPI process of pycbc_live to not wait for the GraceDB event upload process to finish before continuing to process data.
The setup_optimize_snr function depends on the gid returned from the GraceDB event upload interation, so should be started after the upload is finished. As such, it has also been moved to be within the new upload_in_thread function.