Change invspec default truncation to Hann
`inverse_spectrum_truncation` uses boxcar truncation by default. This works OK in the tutorials because the data isn't very long, but it gives a bad filter when used to whiten longer blocks of data. Hann window would be a more sensible default. But we should check whether there's anywhere in the conditioning that intentionally uses the hard truncation.