Change the option name of —autogating-psd-duration to —autogating-psd-segment-length.
Add --autogating-duration option for PyCBC Live run script to set the amount of data we want to apply autogating on.
The default value is set to 16 s.

Move variables « autogating_duration_length » and « autogating_start_sample » inside the « if self.autogating_threshold it not None » block since those variables are only used inside that block.