Make autogating work for the PyCBC Live early-warning search (#4298)
* Apply autogating on the last 16 s of the buffer instead of the analysis chunk only

Add 2 command lines to the PyCBC Live run script to set the length and overlap in seconds of the segments used to estimate the PSD with Welchs method

* Fix help messages of --autogating-psd-duration and --autogating-psd-stride

* Set autogating_psd_duration and autogating_psd_stride in __init__ function of StrainBuffer class

* Forgot to change detect_loud_glitches(strain) to detect_loud_glitches(self.strain)

* Change --autogating-psd-duration and --autogating-psd-stride from type=int to type=float

* Add documentation of the new kwargs autogating_psd_duration and autogating_psd_stride

* Change the option name of —autogating-psd-duration to —autogating-psd-segment-length.

Add --autogating-duration option for PyCBC Live run script to set the amount of data we want to apply autogating on.
The default value is set to 16 s.

Move variables « autogating_duration_length » and « autogating_start_sample » inside the « if self.autogating_threshold it not None » block since those variables are only used inside that block.

* Add missing ,

* Fix typo

* Change comment message since the command --autogating-duration has been added and 16 s is not a hardcoded value anymore.